requirements.txt
streamlit==1.29.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
snowflake-connector-python==3.6.0
cryptography==41.0.8
pandas==2.1.4
python-dotenv==1.0.0